完成文件结构设计
