+++ lib/storage/Factory.js
@@ -1,0 +1,13 @@
+/**
+ * Created by vt on 15/10/8.
+ */
+
+'use strict';
+
+var StorageFactory = module.exports = function () {
+
+};
+
+StorageFactory.create = function () {
+
+};
