+++ lib/storage/memory-storage.js
@@ -1,0 +1,6 @@
+/**
+ * Created by vt on 15/10/8.
+ */
+
+'use strict';
+
